+++ testlib/src/test/java/com/diffplug/spotless/generic/LicenseHeaderStepTest.java
@@ -33,7 +33,7 @@ import com.diffplug.spotless.generic.LicenseHeaderStep.YearMode;
 
 class LicenseHeaderStepTest extends ResourceHarness {
 	private static final String FILE_NO_LICENSE = "license/FileWithoutLicenseHeader.test";
-	private static final String package_ = "package ";
+	private static final String package_ = LicenseHeaderStep.DEFAULT_JAVA_HEADER_DELIMITER;
 	private static final String HEADER_WITH_$YEAR = "This is a fake license, $YEAR. ACME corp.";
 	private static final String HEADER_WITH_RANGE_TO_$YEAR = "This is a fake license with range, 2009-$YEAR. ACME corp.";
 	private static final String HEADER_WITH_$FILE = "This is a fake license, $FILE. ACME corp.";
@@ -286,12 +286,29 @@ class LicenseHeaderStepTest extends ResourceHarness {
 	void should_apply_license_containing_YEAR_filename_token() throws Exception {
 		FormatterStep step = LicenseHeaderStep.headerDelimiter(header(HEADER_WITH_$YEAR_$FILE), package_).build();
 		StepHarnessWithFile.forStep(this, step)
-				.test(
-						new File("Test.java"),
-						getTestResource(FILE_NO_LICENSE),
-						hasHeaderYearFileName(HEADER_WITH_$YEAR_$FILE, currentYear(), "Test.java"))
-				.testUnaffected(
-						new File("Test.java"),
-						hasHeaderYearFileName(HEADER_WITH_$YEAR_$FILE, currentYear(), "Test.java"));
+			.test(
+				new File("Test.java"),
+				getTestResource(FILE_NO_LICENSE),
+				hasHeaderYearFileName(HEADER_WITH_$YEAR_$FILE, currentYear(), "Test.java"))
+			.testUnaffected(
+				new File("Test.java"),
+				hasHeaderYearFileName(HEADER_WITH_$YEAR_$FILE, currentYear(), "Test.java"));
+	}
+
+	void noPackage() throws Throwable {
+		String header = header(getTestResource("license/TestLicense"));
+		FormatterStep step = LicenseHeaderStep.headerDelimiter(header, package_).build();
+		StepHarness.forStep(step)
+				.test(ResourceHarness.getTestResource("license/HelloWorld_java.test"), header + ResourceHarness.getTestResource("license/HelloWorld_java.test"))
+				.test(ResourceHarness.getTestResource("license/HelloWorld_withImport_java.test"), header + ResourceHarness.getTestResource("license/HelloWorld_withImport_java.test"));
+	}
+
+	// The following demonstrate the use of 'module' keyword
+	@Test
+	void moduleInfo() throws Throwable {
+		String header = header(getTestResource("license/TestLicense"));
+		FormatterStep step = LicenseHeaderStep.headerDelimiter(header, package_).build();
+		StepHarness.forStep(step)
+				.test(ResourceHarness.getTestResource("license/module-info.test"), header + ResourceHarness.getTestResource("license/module-info.test"));
 	}
 }